+++ platform/src/main/java/org/stellar/anchor/platform/AnchorPlatformServer.java
@@ -15,6 +15,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.stellar.anchor.config.EventConfig;
 import org.stellar.anchor.event.EventService;
+import org.stellar.anchor.event.KafkaEventService;
 import org.stellar.anchor.platform.configurator.DataAccessConfigurator;
 import org.stellar.anchor.platform.configurator.PlatformAppConfigurator;
 import org.stellar.anchor.platform.configurator.PropertiesReader;
@@ -67,6 +68,6 @@ public class AnchorPlatformServer implements WebMvcConfigurer {
 
   @Bean
   public EventService eventService(EventConfig eventConfig) {
-    return new EventService(eventConfig);
+    return new KafkaEventService(eventConfig);
   }
 }
