+++ platform/src/test/resources/test-anchor-config.yaml
@@ -77,6 +77,17 @@ app-config:
       horizonUrl: https://horizon-testnet.stellar.org
       secretKey: secret # stellar account secret key
 
+  event:
+    enabled: true
+    kafkaBootstrapServer: localhost:29092
+    useSingleQueue: false
+    queues:
+      all: ap_event_single_queue
+      quote_created: ap_event_quote_created
+      transaction_created: ap_event_transaction_created
+      transaction_payment_received: ap_event_transaction_payment_received
+      transaction_error: ap_event_transaction_error
+
 #
 # Spring Data JDBC settings
 #
